+++ quiz/__about__.py
@@ -1,5 +1,5 @@
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 __author__ = "Arie Bovenberg"
-__copyright__ = "2018, Arie Bovenberg"
+__copyright__ = "2019, Arie Bovenberg"
 __license__ = 'Apache 2.0'
 __description__ = 'Capable GraphQL client'